Fix Preconditions.checkState() assertion
According to the TS docs, the form `asserts condition is ...` only narrows the type of the given variable `condition` to the defined type. The type of other variables used to create `condition` is not narrowed.

This kind of assertion needs to use the short form instead (without `is ...`). This ensures that the given boolean condition evaluates true for the rest of the scope, which can narrow the type of variables used within the condition.

https://www.typescriptlang.org/docs/handbook/release-notes/typescript-3-7.html#assertion-functions